+++ components/zcash_address/src/kind/p2pkh.rs
@@ -1,5 +1,0 @@
-/// The prefix for a Base58Check-encoded mainnet transparent P2PKH address.
-pub(crate) const MAINNET: [u8; 2] = [0x1c, 0xb8];
-
-/// The prefix for a Base58Check-encoded testnet transparent P2PKH address.
-pub(crate) const TESTNET: [u8; 2] = [0x1d, 0x25];