+++ zcash_primitives/src/legacy/keys.rs
@@ -7,9 +7,10 @@ use hdwallet::{
 use secp256k1::PublicKey;
 use sha2::{Digest, Sha256};
 use subtle::{Choice, ConstantTimeEq};
-use zcash_spec::PrfExpand;
 
-use crate::{consensus, zip32::AccountId};
+use zcash_protocol::consensus::{self, NetworkConstants};
+use zcash_spec::PrfExpand;
+use zip32::AccountId;
 
 use super::TransparentAddress;
 
@@ -119,7 +120,9 @@ impl AccountPrivKey {
     ) -> Result<AccountPrivKey, hdwallet::error::Error> {
         ExtendedPrivKey::with_seed(seed)?
             .derive_private_key(KeyIndex::hardened_from_normalize_index(44)?)?
-            .derive_private_key(KeyIndex::hardened_from_normalize_index(params.coin_type())?)?
+            .derive_private_key(KeyIndex::hardened_from_normalize_index(
+                params.network_type().coin_type(),
+            )?)?
             .derive_private_key(KeyIndex::hardened_from_normalize_index(account.into())?)
             .map(AccountPrivKey)
     }